+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 - Remove `ShortCodeResultsScreen`
 - Deprecate `PatientRepository#search`
 - Remove RxBinding2
+- Use new method of SQL performance profiling (transform generated Room DAOs)
 
 ### Features
 - Add NHID support in `PatientEntryScreen`